requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
praw>=7.7.1
kafka-python>=2.0.2
happybase>=1.2.0
pyspark>=3.4.1
transformers>=4.30.2
torch>=2.0.1
streamlit>=1.24.0
pandas>=2.0.3
numpy>=1.24.3
setuptools>=65.5.1
wheel>=0.38.4
